# backend/docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateCommentRequest:
    properties:
      book_name:
        maxLength: 200
        type: string
      chapter_name:
        maxLength: 200
        type: string
      chapter_url:
        maxLength: 2048
        type: string
      comment_key:
        maxLength: 200
        type: string
      content:
        maxLength: 1000
        minLength: 1
        type: string
    required:
    - comment_key
    - content
    type: object
...
  model.MigrateCommentKeyRequest:
    properties:
      new_key:
        maxLength: 200
        type: string
      old_key:
        maxLength: 200
        type: string
    required:
    - new_key
    type: object
...
  model.Response:
    properties:
      code:
        description: 五位业务码，"00000" 表示成功
        type: string
      data: {}
      error:
        description: 业务错误文案（成功时为空串）
        type: string
    type: object
...
info:
  contact: {}
  description: 电子书阅读器后端 API（账号/认证/评论/操作日志）。业务响应用统一信封 {code,error,data}，HTTP 恒为 200；需认证的接口用
    Bearer access token。
  title: ebook-server API
  version: 1.0.0
paths:
...
  /api/comments:
    get:
      description: |-
        获取评论列表。comment_keys 是 M2 聚合键过滤（逗号分隔，返回并集）；
        chapter_url / book_name 已废弃，仅为让未换键的历史行继续可读而保留。
      parameters:
      - description: M2 聚合键列表（逗号分隔，返回并集；最多 50 个，单键最长 200）
        in: query
        name: comment_keys
        type: string
      - collectionFormat: multi
        description: 已废弃：旧聚合键（可传多个，返回并集；最多 50 个，单键最长 2048）
        in: query
        items:
          type: string
        name: chapter_url
        type: array
      - description: 已废弃：书名（配合 chapter_url 二次过滤，或单独过滤全书）
        in: query
        name: book_name
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取评论列表
      tags:
      - 评论
    post:
      consumes:
      - application/json
      description: |-
        创建新评论。comment_key（M2 聚合键）必填，服务端只存不解释其格式；
        chapter_url / chapter_name / book_name 可选，仅作过渡期兼容与展示快照。
      parameters:
      - description: 评论内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建评论
      tags:
      - 评论
  /api/comments/{id}:
    delete:
      description: 删除指定评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除评论
      tags:
      - 评论
  /api/comments/migrate:
    post:
      consumes:
      - application/json
      description: |-
        将当前用户在旧聚合键（comment_key）下的评论批量迁移到新聚合键（合并书籍 / 改元数据修键场景）
        new_key 必填非空——迁往空键会让评论落进任何 comment_keys 查询都命中不到的桶。
        old_key 可为空串：空 = 尚未换键的历史行，因此「把本人旧评论收进正确桶」这一方向可表达；单键最长 200。
        只影响本人评论；新旧相同返回 A0305；无匹配评论返回 0（幂等，可重复调用）。
      parameters:
      - description: 迁移请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.MigrateCommentKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 迁移评论聚合键
      tags:
      - 评论
  /api/comments/my:
    get:
      description: 获取当前用户的评论列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取我的评论列表
      tags:
      - 评论
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"